Add azure, vertex, bedrock integrations
