+++ example-apps/workplace-search/api/elasticsearch_client.py
@@ -1,0 +1,16 @@
+from elasticsearch import Elasticsearch
+from lib.elasticsearch_chat_message_history import ElasticsearchChatMessageHistory
+import os
+
+ELASTIC_CLOUD_ID = os.getenv("ELASTIC_CLOUD_ID")
+ELASTIC_USERNAME = os.getenv("ELASTIC_USERNAME")
+ELASTIC_PASSWORD = os.getenv("ELASTIC_PASSWORD")
+
+elasticsearch_client = Elasticsearch(
+    cloud_id=ELASTIC_CLOUD_ID, basic_auth=(ELASTIC_USERNAME, ELASTIC_PASSWORD)
+)
+
+def get_elasticsearch_chat_message_history(index, session_id):
+    return ElasticsearchChatMessageHistory(
+        client=elasticsearch_client, index=index, session_id=session_id
+    )
