+++ example-apps/workplace-search/requirements.txt
@@ -11,13 +11,16 @@ dataclasses-json==0.5.14
 elastic-transport==8.4.0
 elasticsearch==8.9.0
 Flask==2.3.3
+Flask-Cors==4.0.0
 frozenlist==1.4.0
+google-auth==2.23.0
 idna==3.4
 itsdangerous==2.1.2
 Jinja2==3.1.2
+jmespath==1.0.1
 jq==1.4.1
-langchain==0.0.271
-langsmith==0.0.26
+langchain==0.0.297
+langsmith==0.0.39
 MarkupSafe==2.1.3
 marshmallow==3.20.1
 multidict==6.0.4
@@ -26,10 +29,16 @@ numexpr==2.8.5
 numpy==1.25.2
 openai==0.27.9
 packaging==23.1
+pyasn1==0.5.0
+pyasn1-modules==0.3.0
 pydantic==2.3.0
 pydantic_core==2.6.3
+python-dateutil==2.8.2
 PyYAML==6.0.1
 requests==2.31.0
+rsa==4.9
+s3transfer==0.6.2
+six==1.16.0
 SQLAlchemy==2.0.20
 tenacity==8.2.3
 tqdm==4.66.1
@@ -38,4 +47,3 @@ typing_extensions==4.7.1
 urllib3==1.26.16
 Werkzeug==2.3.7
 yarl==1.9.2
-flask-cors
